refactor: 검색 관련 도메인 패키징 및 불필요한 정적 팩토리 메소드 제거
